+++ src/utils/proxy.js
@@ -24,7 +24,7 @@ function sanitize(str) {
 
 /**
  * ObservableSlim
- * @type {{create, observe, pause, resume, remove, beforeChange}}
+ * @type {{create, observe, remove, beforeChange, beginRender, endRender}}
  */
 const ObservableSlim = (function () {
     // An array that stores all of the observables created through the public create() method below.
@@ -41,6 +41,8 @@ const ObservableSlim = (function () {
     // however, in order to avoid an infinite loop of Proxies triggering and re-triggering one another, we use dupProxy
     // to track that a given Proxy was modified from the 'set' handler
     let dupProxy = null;
+
+    let _instance = null;
 
     let _getProperty = function (obj, path) {
         return path.split('.').reduce(function (prev, curr) {
@@ -77,9 +79,6 @@ const ObservableSlim = (function () {
         let calls = 0;
 
         let _notifyObservers = function (numChanges) {
-
-            // if the observable is paused, then we don't want to execute any of the observer functions
-            if (observable.paused === true) return;
 
             // reset calls number after 10ms
             if (autoDomDelay) {
@@ -250,6 +249,16 @@ const ObservableSlim = (function () {
                     let type = 'update';
                     if (typeOfTargetProp === 'undefined') type = 'add';
 
+                    if(_instance.propsComputed) {
+                        if (typeof _instance.propsComputed === 'object') {
+                            const propPath = _instance.propsComputed[currentPath];
+                            const func = _instance[propPath] || propPath;
+                            if (typeof func === 'function') {
+                                value = func.call(_instance, value, receiver[property])
+                            }
+                        }
+                    }
+
                     // store the change that just occurred. it is important that we store the change before invoking the other proxies so that the previousValue is correct
                     changes.push({
                         type: type,
@@ -405,7 +414,6 @@ const ObservableSlim = (function () {
                 observers: [],
                 targets: [target],
                 proxies: [proxy],
-                paused: false,
                 path: path
             };
             observables.push(observable);
@@ -447,9 +455,12 @@ const ObservableSlim = (function () {
          * @param target {Object} required, plain JavaScript object that we want to observe for changes.
          * @param domDelay {Boolean} if true, then batch up changes on a 10ms delay so a series of changes can be processed in one DOM update.
          * @param observer {Function} optional, will be invoked when a change is made to the proxy.
+         * @param instance
          * @returns {Object}
          */
-        create: function (target, domDelay, observer) {
+        create: function (target, domDelay, observer, instance) {
+
+            _instance = instance;
 
             // test if the target is a Proxy, if it is then we need to retrieve the original object behind the Proxy.
             // we do not allow creating proxies of proxies because -- given the recursive design of ObservableSlim -- it would lead to sharp increases in memory usage